fix(migration): reconcile gamification level-sync trigger to match prod exactly
The deployed private.sync_user_level_from_xp() casts SUM(amount) to ::integer
before assigning to the integer v_total; the committed file omitted the explicit
cast (relying on the implicit assignment cast — functionally identical). Verified
via the prod migration ledger (statements column): adding ::integer makes the
committed file byte-faithful to what was applied. No behavior change; a fresh
db reset now reproduces prod exactly for this function.

+++ supabase/migrations/20260709113822_gamification_atomic_level_sync.sql
@@ -27,7 +27,10 @@ DECLARE
   v_title text;
 BEGIN
   -- Authoritative total: recompute from the ledger, never trust a delta.
-  SELECT COALESCE(SUM(amount), 0) INTO v_total
+  -- Cast SUM (bigint) to integer to match the one-time heal below and the
+  -- deployed prod function exactly (v_total is integer; the cast is explicit
+  -- rather than relying on the implicit assignment cast).
+  SELECT COALESCE(SUM(amount), 0)::integer INTO v_total
   FROM public.xp_transactions
   WHERE user_id = NEW.user_id;
